requirements.txt
pydantic>=2.7.0
pydantic-settings>=2.2.0
azure-storage-blob>=12.19.0
httpx>=0.27.0
